Added support for padding
In this version update (2022.11.09) the FUME image translation layer
enables the user to translate images of irregular sizes (e.g. padding).
To do so, a central convention was changed. The fundamental matrix was
mapping indices onto indices so far. Now, to support scaling, the
indices are converted to array-centered coordinates first.

This enables:
* padding and downsampling of images
* images sizes which are not a downsampled version of 976